Move common subsampling logic to filter_support
These can be used by any engine.
