+++ augur/filter.py
@@ -14,10 +14,10 @@ import pandas as pd
 import random
 import re
 from tempfile import NamedTemporaryFile
-from typing import Collection
 
 from augur.filter_support.exceptions import FilterException
 from augur.filter_support.output import filter_kwargs_to_str
+from augur.filter_support.subsample import TooManyGroupsError, calculate_sequences_per_group
 
 from .dates import numeric_date, numeric_date_type, SUPPORTED_DATE_HELP_TEXT, is_date_ambiguous, get_numerical_dates
 from .errors import AugurError
@@ -1640,181 +1640,6 @@ def run(args):
         raise AugurError("All samples have been dropped! Check filter rules and metadata file format.")
 
     print(f"{total_strains_passed} strains passed all filters")
-
-
-def calculate_sequences_per_group(target_max_value, counts_per_group, allow_probabilistic=True):
-    """Calculate the number of sequences per group for a given maximum number of
-    sequences to be returned and the number of sequences in each requested
-    group. Optionally, allow the result to be probabilistic such that the mean
-    result of a Poisson process achieves the calculated sequences per group for
-    the given maximum.
-
-    Parameters
-    ----------
-    target_max_value : int
-        Maximum number of sequences to return by subsampling at some calculated
-        number of sequences per group for the given counts per group.
-    counts_per_group : list[int]
-        A list with the number of sequences in each requested group.
-    allow_probabilistic : bool
-        Whether to allow probabilistic subsampling when the number of groups
-        exceeds the requested maximum.
-
-    Raises
-    ------
-    TooManyGroupsError :
-        When there are more groups than sequences per group and probabilistic
-        subsampling is not allowed.
-
-    Returns
-    -------
-    int or float :
-        Number of sequences per group.
-    bool :
-        Whether probabilistic subsampling was used.
-
-    """
-    probabilistic_used = False
-
-    try:
-        sequences_per_group = _calculate_sequences_per_group(
-            target_max_value,
-            counts_per_group,
-        )
-    except TooManyGroupsError as error:
-        if allow_probabilistic:
-            print_err(f"WARNING: {error}")
-            sequences_per_group = _calculate_fractional_sequences_per_group(
-                target_max_value,
-                counts_per_group,
-            )
-            probabilistic_used = True
-        else:
-            raise error
-
-    return sequences_per_group, probabilistic_used
-
-
-class TooManyGroupsError(ValueError):
-    def __init__(self, msg):
-        self.msg = msg
-
-    def __str__(self):
-        return str(self.msg)
-
-
-def _calculate_total_sequences(
-        hypothetical_spg: float, sequence_lengths: Collection[int],
-) -> float:
-    # calculate how many sequences we'd keep given a hypothetical spg.
-    return sum(
-        min(hypothetical_spg, sequence_length)
-        for sequence_length in sequence_lengths
-    )
-
-
-def _calculate_sequences_per_group(
-        target_max_value: int,
-        sequence_lengths: Collection[int]
-) -> int:
-    """This is partially inspired by
-    https://github.com/python/cpython/blob/3.8/Lib/bisect.py
-
-    This should return the spg such that we don't exceed the requested
-    number of samples.
-
-    Parameters
-    ----------
-    target_max_value : int
-        the total number of sequences allowed across all groups
-    sequence_lengths : Collection[int]
-        the number of sequences in each group
-
-    Returns
-    -------
-    int
-        maximum number of sequences allowed per group to meet the required maximum total
-        sequences allowed
-
-    >>> _calculate_sequences_per_group(4, [4, 2])
-    2
-    >>> _calculate_sequences_per_group(2, [4, 2])
-    1
-    >>> _calculate_sequences_per_group(1, [4, 2])
-    Traceback (most recent call last):
-        ...
-    augur.filter.TooManyGroupsError: Asked to provide at most 1 sequences, but there are 2 groups.
-    """
-
-    if len(sequence_lengths) > target_max_value:
-        # we have more groups than sequences we are allowed, which is an
-        # error.
-
-        raise TooManyGroupsError(
-            "Asked to provide at most {} sequences, but there are {} "
-            "groups.".format(target_max_value, len(sequence_lengths)))
-
-    lo = 1
-    hi = target_max_value
-
-    while hi - lo > 2:
-        mid = (hi + lo) // 2
-        if _calculate_total_sequences(mid, sequence_lengths) <= target_max_value:
-            lo = mid
-        else:
-            hi = mid
-
-    if _calculate_total_sequences(hi, sequence_lengths) <= target_max_value:
-        return int(hi)
-    else:
-        return int(lo)
-
-
-def _calculate_fractional_sequences_per_group(
-        target_max_value: int,
-        sequence_lengths: Collection[int]
-) -> float:
-    """Returns the fractional sequences per group for the given list of group
-    sequences such that the total doesn't exceed the requested number of
-    samples.
-
-    Parameters
-    ----------
-    target_max_value : int
-        the total number of sequences allowed across all groups
-    sequence_lengths : Collection[int]
-        the number of sequences in each group
-
-    Returns
-    -------
-    float
-        fractional maximum number of sequences allowed per group to meet the
-        required maximum total sequences allowed
-
-    >>> np.around(_calculate_fractional_sequences_per_group(4, [4, 2]), 4)
-    1.9375
-    >>> np.around(_calculate_fractional_sequences_per_group(2, [4, 2]), 4)
-    0.9688
-
-    Unlike the integer-based version of this function, the fractional version
-    can accept a maximum number of sequences that exceeds the number of groups.
-    In this case, the function returns a fraction that can be used downstream,
-    for example with Poisson sampling.
-
-    >>> np.around(_calculate_fractional_sequences_per_group(1, [4, 2]), 4)
-    0.4844
-    """
-    lo = 1e-5
-    hi = target_max_value
-
-    while (hi / lo) > 1.1:
-        mid = (lo + hi) / 2
-        if _calculate_total_sequences(mid, sequence_lengths) <= target_max_value:
-            lo = mid
-        else:
-            hi = mid
-
-    return (lo + hi) / 2
 
 
 def _cleanup_outputs(args):
